Refactor BME680 gas sensor blocks to new module.
