+++ src/modules/blockly/generators/propc/sensors.js
@@ -62,7 +62,7 @@ Blockly.Blocks.sound_impact_run = {
     this.setNextStatement(true, null);
     this.setPreviousStatement(true, 'Block');
 
-    // Prepare the Pn dropdown list
+    // Prepare the Pin dropdown list
     this.userDefinedConstantsList_ = buildConstantsList();
     this.setPinMenus();
   },
@@ -262,500 +262,3 @@ Blockly.propc.sound_impact_end = function() {
   }
 };
 
-// //
-// //
-// //
-// //
-// // ------------------ 4x4 Keypad Blocks ----------------------------------------
-// //
-// //
-// //
-// //
-//
-// /**
-//  * 4x4 Keypad Initialization block
-//  */
-// Blockly.Blocks.keypad_initialize = {
-//   helpUrl: Blockly.MSG_KEYPAD_HELPURL,
-//
-//   init: function() {
-//     this.setTooltip(Blockly.MSG_KEYPAD_INIT_TOOLTIP);
-//     this.setColour(colorPalette.getColor('input'));
-//     this.appendDummyInput('PINS');
-//     this.setInputsInline(true);
-//     this.setPreviousStatement(true, 'Block');
-//     this.setNextStatement(true, null);
-//
-//     // Prepare the Pn dropdown list
-//     this.userDefinedConstantsList_ = buildConstantsList();
-//     this.setPinMenus();
-//   },
-//
-//   /**
-//    * This is a called from within the updateConstMenu method.
-//    *
-//    * @description This method is called as a finalizer in the updateConstMenu() call. This
-//    * call is by reference to the method defined in the 'sound_impact_run' block. That method
-//    * calls setPinMenus with no parameters, which basically renders this method useless.
-//    *
-//    * @param {string} oldValue
-//    * @param {string} newValue
-//    */
-//   setPinMenus: function(oldValue = '', newValue = '') {
-//     const profile = getDefaultProfile();
-//     const m = [];
-//     for (let i = 0; i < 8; i++) {
-//       m[i] = this.getFieldValue('P' + i.toString(10));
-//     }
-//     if (this.getInput('PINS')) {
-//       this.removeInput('PINS');
-//     }
-//     const pinConstantList = this.userDefinedConstantsList_.map(function(value) {
-//       return [value, value];
-//     });
-//     this.appendDummyInput('PINS')
-//         .appendField('4x4 Keypad initialize PINS left')
-//         .appendField(new Blockly.FieldDropdown(
-//             profile.digital.concat(pinConstantList)), 'P0')
-//         .appendField(new Blockly.FieldDropdown(
-//             profile.digital.concat(pinConstantList)), 'P1')
-//         .appendField(new Blockly.FieldDropdown(
-//             profile.digital.concat(pinConstantList)), 'P2')
-//         .appendField(new Blockly.FieldDropdown(
-//             profile.digital.concat(pinConstantList)), 'P3')
-//         .appendField('|')
-//         .appendField(new Blockly.FieldDropdown(
-//             profile.digital.concat(pinConstantList)), 'P4')
-//         .appendField(new Blockly.FieldDropdown(
-//             profile.digital.concat(pinConstantList)), 'P5')
-//         .appendField(new Blockly.FieldDropdown(
-//             profile.digital.concat(pinConstantList)), 'P6')
-//         .appendField(new Blockly.FieldDropdown(
-//             profile.digital.concat(pinConstantList)), 'P7')
-//         .appendField('right');
-//     for (let i = 0; i < 8; i++) {
-//       if (m[i] && m[i] === oldValue && newValue) {
-//         this.setFieldValue(newValue, 'P' + i.toString(10));
-//       } else if (m[i]) {
-//         this.setFieldValue(m[i], 'P' + i.toString(10));
-//       }
-//     }
-//   },
-// };
-//
-// /**
-//  * Keypad Initialization C code generator
-//  * @return {string}
-//  */
-// Blockly.propc.keypad_initialize = function() {
-//   if (!this.disabled) {
-//     const profile = getDefaultProfile();
-//     const kp = [];
-//     for (let k = 0; k < 8; k++) {
-//       kp[k] = this.getFieldValue('P' + k);
-//       if (profile.digital.toString().indexOf(kp[k] + ',' + kp[k]) === -1) {
-//         kp[k] = 'MY_' + kp[k];
-//       }
-//     }
-//     let keypadVars = 'int __rowPins[] = {' + kp[0] + ', ' + kp[1] + ', ' +
-//         kp[2] + ', ' + kp[3] + '};\n';
-//     keypadVars += 'int __colPins[] = {' + kp[4] + ', ' + kp[5] + ', ' +
-//         kp[6] + ', ' + kp[7] + '};\n';
-//     keypadVars += 'int __buttonVals[] = {1, 2, 3, \'A\', 4, 5, 6, \'B\', 7,' +
-//         ' 8, 9, \'C\', \'*\', 0, \'#\', \'D\'};\n';
-//     Blockly.propc.definitions_['keypad_lib'] = '#include "keypad.h"';
-//     Blockly.propc.global_vars_['keypad_pins'] = keypadVars;
-//     Blockly.propc.setups_['keypad_init'] =
-//         'keypad_setup(4, 4, __rowPins, __colPins, __buttonVals);';
-//   }
-//   return '';
-// };
-//
-// /**
-//  * Scan the keyboard
-//  */
-// Blockly.Blocks.keypad_read = {
-//   helpUrl: Blockly.MSG_KEYPAD_HELPURL,
-//
-//   init: function() {
-//     this.setTooltip(Blockly.MSG_KEYPAD_READ_TOOLTIP);
-//     this.setColour(colorPalette.getColor('input'));
-//     this.appendDummyInput()
-//         .appendField('4x4 Keypad');
-//     this.setOutput(true, null);
-//   },
-//
-//   onchange: function() {
-//     if (!verifyBlockTypeEnabled('keypad_initialize')) {
-//       this.setWarningText('WARNING: You must use a keypad ' +
-//           ' \ninitialize block at the beginning of your program!');
-//     } else {
-//       this.setWarningText(null);
-//     }
-//   },
-// };
-//
-// /**
-//  *
-//  * @return {string|[string, number]}
-//  */
-// Blockly.propc.keypad_read = function() {
-//   if (!verifyBlockTypeEnabled('keypad_initialize')) {
-//     return '// ERROR: Missing Keypad initialize block!';
-//   } else {
-//     return ['keypad_read()', Blockly.propc.ORDER_ATOMIC];
-//   }
-// };
-
-//
-//
-//
-//
-// ------------------ BME680 Air Quality Sensor -----------------------------
-//
-//
-//
-//
-
-/**
- * BME680 gas sensor Initialization
- *
- * @description This block sets the pin assignments for the sensor and calls an
- * SPI init function in the library.
- *
- * NOTE: The block is a SINGLETON but is not implemented as one. If there are
- * multiple init blocks in the project, the last one declared will be the one
- * that provides the pin settings.
- *
- * @type {{
- *  init: Blockly.Blocks.bme680_init.init,
- *  setPinMenus: Blockly.Blocks.bme680_init.setPinMenus,
- *  helpUrl: string,
- *  updateConstMenu: *
- *  }}
- */
-Blockly.Blocks.bme680_init = {
-  helpUrl: Blockly.MSG_BME680_HELPURL,
-
-  init: function() {
-    this.setTooltip(Blockly.MSG_BME680_INIT_TOOLTIP);
-    this.setColour(colorPalette.getColor('input'));
-    this.appendDummyInput('PINS');
-    this.setInputsInline(false);
-    this.setNextStatement(true, null);
-    this.setPreviousStatement(true, 'Block');
-    this.updateConstMenu();
-  },
-
-  /**
-   * Update the Constants menu
-   *
-   * This currently uses the updateConstMenu method in an unrelated block and should
-   * be refactored to remove the dependency.
-   */
-  updateConstMenu: Blockly.Blocks['sound_impact_run'].updateConstMenu,
-
-  /**
-   * Set the pin dropdown menu elements
-   *
-   * @param {string} oldValue
-   * @param {string} newValue
-   */
-  setPinMenus: function(oldValue, newValue) {
-    const profile = getDefaultProfile();
-    const mv = [
-      'PIN_CLK',
-      'PIN_SDI',
-      'PIN_SDO',
-      'PIN_CS',
-    ];
-    const m = [
-      this.getFieldValue('PIN_CLK'),
-      this.getFieldValue('PIN_SDI'),
-      this.getFieldValue('PIN_SDO'),
-      this.getFieldValue('PIN_CS'),
-    ];
-
-    if (this.getInput('PINS')) {
-      this.removeInput('PINS');
-    }
-
-    const pinConstantList = this.userDefinedConstantsList_.map(function(value) {
-      return [value, value];
-    });
-
-    this.appendDummyInput('PINS')
-        .appendField('Air Quality initialize SCK')
-        .appendField(new Blockly.FieldDropdown(
-            profile.digital.concat(pinConstantList)), 'PIN_CLK')
-        .appendField('SDI')
-        .appendField(new Blockly.FieldDropdown(
-            profile.digital.concat(pinConstantList)), 'PIN_SDI')
-        .appendField('SDO')
-        .appendField(new Blockly.FieldDropdown(
-            profile.digital.concat(pinConstantList)), 'PIN_SDO')
-        .appendField('CS')
-        .appendField(new Blockly.FieldDropdown(
-            profile.digital.concat(pinConstantList)), 'PIN_CS');
-
-    // Loop through the array of pin fields
-    for (let i = 0; i < 4; i++) {
-      // If the pin has a value, and the value is equal to the old value, and there is
-      // a new value, then set the pin field to the new value
-      if (m[i] && m[i] === oldValue && newValue) {
-        this.setFieldValue(newValue, mv[i]);
-      } else if (m[i]) {
-        this.setFieldValue(m[i], mv[i]);
-      }
-    }
-  },
-};
-
-/**
- * BME-680 Initialization C code generator
- * @return {string}
- */
-Blockly.propc.bme680_init = function() {
-  const profile = getDefaultProfile();
-  const pin = [
-    this.getFieldValue('PIN_CLK'),
-    this.getFieldValue('PIN_SDI'),
-    this.getFieldValue('PIN_SDO'),
-    this.getFieldValue('PIN_CS'),
-  ];
-  for (let i = 0; i < 3; i++) {
-    if (profile.digital.toString().indexOf(pin[i] + ',' + pin[i]) === -1) {
-      pin[i] = 'MY_' + pin[i];
-    }
-  }
-  if (!this.disabled) {
-    Blockly.propc.definitions_['include_bme680'] = '#include "bme680.h"';
-    Blockly.propc.setups_['init_bme680'] =
-        'gas_sensor = bme680_openSPI(' + pin.join(', ') + ');\n';
-    Blockly.propc.global_vars_['device_bme680'] = 'bme680 *gas_sensor;\n';
-  }
-  return '';
-};
-
-/**
- *
- * @type {{
- *  init: Blockly.Blocks.bme680_read.init,
- *  helpUrl: string,
- *  onchange: Blockly.Blocks.bme680_read.onchange
- *  }}
- */
-Blockly.Blocks.bme680_read = {
-  helpUrl: Blockly.MSG_BME680_HELPURL,
-  init: function() {
-    this.setTooltip(Blockly.MSG_BME680_READ_TOOLTIP);
-    this.setColour(colorPalette.getColor('input'));
-    this.appendDummyInput()
-        .appendField('Air Quality read');
-    this.setInputsInline(false);
-    this.setNextStatement(true, null);
-    this.setPreviousStatement(true, 'Block');
-  },
-
-  onchange: function() {
-    const allBlocks = Blockly.getMainWorkspace().getAllBlocks().toString();
-    if (allBlocks.indexOf('Air Quality initialize') === -1) {
-      this.setWarningText('WARNING: You must use an Air Quality\ninitialize' +
-          ' block at the beginning of your program!');
-    } else {
-      this.setWarningText(null);
-    }
-  },
-};
-
-/**
- *
- * @return {string}
- */
-Blockly.propc.bme680_read = function() {
-  let code = '';
-  const allBlocks = Blockly.getMainWorkspace().getAllBlocks().toString();
-  if (allBlocks.indexOf('Air Quality initialize') === -1) {
-    code += '// ERROR: Missing Air Quality initialize block!';
-  } else {
-    code += 'bme680_readSensor(gas_sensor);';
-  }
-  return code;
-};
-
-/**
- *
- * @type {{
- *  init: Blockly.Blocks.bme680_heater.init,
- *  helpUrl: string,
- *  onchange: Blockly.Blocks.bme680_heater.onchange
- *  }}
- */
-Blockly.Blocks.bme680_heater = {
-  helpUrl: Blockly.MSG_BME680_HELPURL,
-  init: function() {
-    this.setTooltip(Blockly.MSG_BME680_HEATER);
-    this.setColour(colorPalette.getColor('input'));
-    this.appendDummyInput()
-        .appendField('Air Quality heater')
-        .appendField(new Blockly.FieldDropdown([
-          ['disable', 'Disable'],
-          ['enable', 'Enable'],
-        ]), 'HEAT_STATE');
-    this.setInputsInline(false);
-    this.setNextStatement(true, null);
-    this.setPreviousStatement(true, 'Block');
-  },
-
-  onchange: function() {
-    const allBlocks = Blockly.getMainWorkspace().getAllBlocks().toString();
-    if (allBlocks.indexOf('Air Quality initialize') === -1) {
-      this.setWarningText('WARNING: You must use an Air Quality\ninitialize' +
-          ' block at the beginning of your program!');
-    } else {
-      this.setWarningText(null);
-    }
-  },
-};
-
-/**
- *
- * @return {string}
- */
-Blockly.propc.bme680_heater = function() {
-  let code = '';
-  const allBlocks = Blockly.getMainWorkspace().getAllBlocks().toString();
-  if (allBlocks.indexOf('Air Quality initialize') === -1) {
-    code += '// ERROR: Missing Air Quality initialize block!';
-  } else {
-    code += 'bme680_heater' + this.getFieldValue(
-        'HEAT_STATE') + '(gas_sensor);';
-  }
-  return code;
-};
-
-/**
- *
- * @type {{
- *  init: Blockly.Blocks.bme680_get_value.init,
- *  mutationToDom: (function(): HTMLElement),
- *  helpUrl: string,
- *  onchange: Blockly.Blocks.bme680_get_value.onchange,
- *  setMeasUnit: Blockly.Blocks.bme680_get_value.setMeasUnit,
- *  domToMutation: Blockly.Blocks.bme680_get_value.domToMutation
- *  }}
- */
-Blockly.Blocks.bme680_get_value = {
-  helpUrl: Blockly.MSG_BME680_HELPURL,
-  init: function() {
-    this.measUnits = {
-      'temperature': [
-        ['in \u00b0C', 'CELSIUS'],
-        ['in \u00b0F', 'FAHRENHEIT'],
-        ['in Kelvin', 'KELVIN'],
-      ],
-      'pressure': [
-        ['in pascals', 'PASCALS'],
-        ['in mmHg', 'MMHG'],
-        ['in inHg', 'INHG'],
-        ['in PSI', 'PSI'],
-      ],
-      'humidity': [['in %', ' ']],
-      'gasResistance': [['in \u2126', ' ']],
-      'altitude': [
-        ['in meters', 'METERS'],
-        ['in feet', 'FEET'],
-      ],
-    };
-    this.setTooltip(Blockly.MSG_BME680_GET_VALUE_TOOLTIP);
-    this.setColour(colorPalette.getColor('input'));
-    this.appendDummyInput()
-        .appendField('Air Quality')
-        .appendField(new Blockly.FieldDropdown([
-          ['Temperature', 'temperature'],
-          ['Pressure', 'pressure'],
-          ['Altitude Estimate', 'altitude'],
-          ['Gas Sensor Resistance', 'gasResistance'],
-          ['Relative Humidity', 'humidity'],
-        ],
-        function(val) {
-          // eslint-disable-next-line no-invalid-this
-          this.getSourceBlock().setMeasUnit(val);
-        }),
-        'SENSOR');
-    this.appendDummyInput('UNITS')
-        .appendField(new Blockly.FieldDropdown(
-            this.measUnits['temperature']), 'UNIT');
-    this.appendDummyInput('MULTIPLIER')
-        .appendField(new Blockly.FieldDropdown([
-          ['\u2715 1', ' '],
-          ['\u2715 10', '*10.0'],
-          ['\u2715 100', '*100.0'],
-          ['\u2715 1000', '*1000.0'],
-          ['\u2715 10000', '*10000.0'],
-        ]), 'MULT');
-    this.setInputsInline(true);
-    this.setNextStatement(false, null);
-    this.setPreviousStatement(false, null);
-    this.setOutput(true, 'Number');
-  },
-
-  setMeasUnit: function(val) {
-    this.removeInput('UNITS');
-    if (val === 'humidity') {
-      this.appendDummyInput('UNITS')
-          .appendField('%');
-    } else if (val === 'gasResistance') {
-      this.appendDummyInput('UNITS')
-          .appendField('\u2126');
-    } else {
-      this.appendDummyInput('UNITS')
-          .appendField(new Blockly.FieldDropdown(this.measUnits[val]), 'UNIT');
-    }
-    this.moveInputBefore('UNITS', 'MULTIPLIER');
-  },
-
-  mutationToDom: function() {
-    const container = document.createElement('mutation');
-    container.setAttribute('sensor', this.getFieldValue('SENSOR'));
-    return container;
-  },
-
-  domToMutation: function(container) {
-    const val = container.getAttribute('sensor');
-    if (val) {
-      this.setMeasUnit(val);
-    }
-  },
-
-  onchange: function() {
-    const allBlocks = Blockly.getMainWorkspace().getAllBlocks().toString();
-    if (allBlocks.indexOf('Air Quality initialize') === -1) {
-      this.setWarningText('WARNING: You must use an Air Quality\ninitialize' +
-          ' block at the beginning of your program!');
-    } else {
-      this.setWarningText(null);
-    }
-  },
-};
-
-/**
- *
- * @return {[string, number]}
- */
-Blockly.propc.bme680_get_value = function() {
-  let code = '';
-  const allBlocks = Blockly.getMainWorkspace().getAllBlocks().toString();
-  if (allBlocks.indexOf('Air Quality initialize') === -1) {
-    code += '// ERROR: Missing Air Quality initialize block!';
-  } else {
-    const sensor = this.getFieldValue('SENSOR');
-    const mult = this.getFieldValue('MULT') || '';
-    let unit = this.getFieldValue('UNIT') || '';
-    if (unit.length > 2) {
-      unit = ', ' + unit;
-    }
-    code += '(int)(bme680_' + sensor + '(gas_sensor' + unit + ')' + mult + ')';
-  }
-  return [code, Blockly.propc.ORDER_ATOMIC];
-};
